Fix deleteMessages topic ID type

+++ TMessagesProj/src/main/java/org/telegram/ui/Components/ChatActivityEnterView.java
@@ -8011,7 +8011,7 @@ public class ChatActivityEnterView extends FrameLayout implements
                         null,
                         null,
                         dialog_id,
-                        parentFragment.getTopicId(),
+                        (int) parentFragment.getTopicId(),
                         true,
                         parentFragment.getChatMode()
                 );
